Add advanced proxy config for client npm start
Uses a middleware that redirects /api and /auth to the backend when
running npm start's webpack dev server; not just the ajax fetches but
also the actual redirects.

Tested on local windows machine, need to test on a mac.

TODO documentation for setup in readme and delete all the old files.