Stop button from staying hovered when remounted
Add direct setter to useHover return value.

Buttons that disappear on click should be set to hover state false when
pressed (like the add buttons on search results). This is because
otherwise, when they are made visible again, they retain the hover state
and don't become unhovered unless the mouse returns to them and then
leaves.

This is because the button element is not a component of its own and
does not own its hover state; rather the SearchItem is the component
that owns the state and when the button is unmounted, the SearchItem
itself is not unmounted, so its state persists.