FIX package.json running backend wrong
With the addition of client/src/shared/, tsc was building its files
differently. Instead of putting all the translations of the files in
src/ directly in build/, it put those translations in build/src/ to
preserve the directory structure. (It also put the files in
client/src/shared/ into build/client/src/shared/). The main field in
package.json wasn't working with this new build directory structure.