Fix auth error message on page load; redir instead
Auth error messages (using alert()) now only happen if the page is
already loaded and a 4xx/5xx error happens; if the page was just loaded,
`/api/login` is requested to check if the user is logged in before
making any other requests that could fail if not logged in. If the user
is not logged in, they are redirected immediately to `/login` instead of
being told an error happened.

This is useful for initial load of the page before logging in, and also
if a user who has logged in bookmarks a URL but then gets timed out.

TODO fix the case where the user's access_token got timed out, but their
refresh_token could still be used to keep them logged in instead of
redirecting them to `/login` for them to fully log in again.