Refactor search & playlist tracks API #70 #71
Refactor api types for track search and tracks within
`.../playlists/:id` to have just the data needed by the frontend, and
move some of the processing from the frontend in several places to the
backend in one place.

These were done together so the draft addition track's data can be
mocked in `SearchResults.tsx` `addButtonOnClick` callback. This will now
allow spotify data on removed tracks to be saved when they are removed
for #87.