Fix re-add bug from Array.splice
When there were multiple removed tracks and one of them was re-added,
the other removed tracks disappeared because I assumed that if I leave
off the second parameter to splice it would default to 1. (This behavior
likely varied depending on the position of the re-added track within the
removed tracks array, so maybe it didn't always happen).